Add interpreter instead of a backend
The reasoning for this should be explained further in the README at some point.

The eventual goal for Iku is to be able to generate native binaries while benefitting
from some concurrent runtime with a GC. To accomplish this, with reasonable
interop and codegen, the best longterm solution is to use something like LLVM
or cranelift (preferably the latter), accompanied with a runtime (in Rust maybe)
compiled and statically linked with the programs we generate.

Because of this, the first step towards that is to start with an interpreter,
which can get the semantics of the language correct. After that, we can
replace the interpreter with codegen + runtime, and get an equivalent program.
We also benefit from reusing the same frontend, if we move towards a bytecode interpreter.