Add semicolon insertion
This works by making the lexer aware of which tokens can be followed by implicit
semicolons. When the lexer is aware of this, it checks to see if the whitespace
it skips through contains at least one newline, in which case it emits a semicolon.
Emitting a semicolon makes it no longer want to insert an implicit semicolon, because
an explicit semicolon doesn't yield to implicit semicolons afterwards.