Add addition and subtraction
This adds the + and - operators.

The way subtraction vs negative litterals works is that `2 -2` is parsed
as two integer litterals, since the minus character is immediately followed by a digit.
In order to parse a minus character as subtraction, it needs to be followed by something
that isn't a digit.